+++ connect-sdk/src/main/java/com/finicity/connect/sdk/ConnectWebViewClient.java
@@ -20,7 +20,7 @@ class ConnectWebViewClient extends WebViewClient {
         // Compare finished URL to Connect link. For some reason an extra '/' is getting added
         // before the query string by the time it gets here.
         if (url.equals(connectUrl) || url.equals(connectUrl.replace("?", "/?"))) {
-            eventHandler.onLoaded();
+            eventHandler.onLoad();
             mConnect.startPingTimer();
         }
     }