+++ demo-app/src/main/java/com/finicity/connect/demo/MainActivity.java
@@ -11,12 +11,9 @@ import android.widget.EditText;
 
 import com.finicity.connect.sdk.Connect;
 import com.finicity.connect.sdk.EventHandler;
-import com.finicity.connect.sdk.EventListener;
 
 public class MainActivity extends AppCompatActivity {
     private static final String TAG = "MainActivity";
-    private Button mStartButtonEventListener;
-    private Button mStartButtonEventHandler;
     private EditText mEditConnectUrl;
 
     @Override
@@ -24,17 +21,8 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        // Add click listeners for EventListener and EventHandler buttons
-        this.mStartButtonEventListener = findViewById(R.id.startButtonEventListener);
-
-        mStartButtonEventListener.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                launchActivity(new ConsoleEventListener());
-            }
-        });
-
-        this.mStartButtonEventHandler = findViewById(R.id.startButtonEventHandler);
+        // Add click listener for EventListener
+        Button mStartButtonEventHandler = findViewById(R.id.startButtonEventHandler);
 
         mStartButtonEventHandler.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -44,19 +32,6 @@ public class MainActivity extends AppCompatActivity {
         });
 
         this.mEditConnectUrl = findViewById(R.id.editConnectUrl);
-    }
-
-    private void launchActivity(EventListener listener) {
-        String url = getEditConnectUrl();
-
-        if(url.length() > 0) {
-            // Null out text so we can repeat with new link after Connect Activity closes.
-            mEditConnectUrl.setText("");
-
-            Log.i(TAG, ">>> Launching Connect activity");
-
-            Connect.start(this, url, listener);
-        }
     }
 
     private void launchActivity(EventHandler eventHandler) {
@@ -74,7 +49,6 @@ public class MainActivity extends AppCompatActivity {
 
     private String getEditConnectUrl() {
         String rawUrl = this.mEditConnectUrl.getText().toString();
-
         return rawUrl.replace("localhost:", "10.0.2.2:");
     }
 }